Add files via upload
The intercept is:  93.91995135417807
The coefficients of the variables are not in sync with their correlation coefficients with the target variable. For example, the coefficient of "UNRATE" (unemployment rate) is positive. It does not make sense as a rise in unemployment will surely lead to a decrease in the home prices. So, the question arises, why did this happen?
As most variable increase with time. So, it is expected that they will have high correlation. So, instead of depending upon the linear model, we can directly study the impact of variables on the home price index using the scatter plots.
A few variables that could have been studied are below.
Net-immigration (It is supposed to have a positive impact. No suitable data could be found)
Marriage Rate (People tend to buy homes after they get married. So, it might have some effect. No data could be found)
Average house size (The data was available only for the years after 2015. Though it is expected that an increase in the average house size would increase prices, it is found that the average home size has been consistently decreasing although the prica has been increasing)
Land availability (Less land, higher prices. Perhaps this is why the prices has been rising even though the average house size has been decreasing. No relevant data could be found)
Tax Rate (Too many brackets (7) and could not analyse due to time)
Number of active listings (Data prior to 2017 was not found)